Censorship Removed, Docker files Moved. 1. Censorship part was deleted. (Moved at Legacy_Codes Directory) 2. For the Readability, some docker files were moved to Docker Directory
