+++ config.example.py
@@ -35,8 +35,7 @@ PREFIXES = ["이프야 ", "ㅇ", "잎", "ep "]
 ACTIVITIES = [
     "{}곳의 서버에서 검열 삭제",
     "{}곳의 서버에서 낚시",
-    "'이프야 도움말'을 입력해 보라고 이야기",
-    "EpBot | ep help",
+    "'/도움말'을 입력해 보라고 이야기",
 ]
 DEBUG_ACTIVITIES = ["버그 수정 중. . ."]
